Library

Library

Misc

Spherical harmonics evaluation strategy

Models

Algorithms

Sampling

Plotting

Misc

Tractography.Cone — Type

struct Cone{T<:Real}

Structure to encode a cone to limit sampling the direction. This ensures that the angle in degrees between to consecutive streamline directions is less than angle.

The implemented condition is for cn = Cone(angle)

(cn::Cone)(d1, d2) = dot(d1, d2) > cos(pi/180 * cn.alpha)

Fields

alpha::Real: half section angle in degrees

Constructor

Cone(angle)

Tractography._apply_mask! — Function

_apply_mask!(model, mask)

Multiply the mask which is akin to a matrix of Bool with same size as the data stored in model. Basically, the mask mask[ix,iy,iz] ensures whether the voxel (ix,iy,iz) is discarded or not.

Arguments

model::TMC.

mask can be a AbstractArray{3, Bool} or a NIVolume.

Tractography.save_streamlines — Function

save_streamlines(filename, tractogram, tracto_length)

Save tractogram in filename in a tck format. You should probably use a filename like streamlines-julia.tck.

Arguments

filename::String file name.

tractogram can either be an abstract array of dimension 3 x n_length_fiber x n_fibres or a Vector{ <: TractoResult} after a call to sample.

Notes

The current implementation is based on NiBabel.

Missing docstring.

Missing docstring for Tractography.Exp𝕊². Check Documenter's build log for details.

Tractography.spherical_to_euclidean — Function

spherical_to_euclidean(θ, ϕ)

Transform spherical to cartesian coordinates.

Recall that t∈[0, π] and p∈[0, 2π]

Tractography.euclidean_to_spherical — Function

euclidean_to_spherical(x, y, z)

Transform cartesian to spherical coordinates. Assume that the vector has norm one.

Spherical harmonics evaluation strategy

Tractography.FibonacciSH — Type

struct FibonacciSH <: Tractography.AbstractEvaluation

Spherical harmonics evaluation based on Fibonacci sampling. The ODF are computed on the fly using matrix-vector. Requires little memory.

See also ComputeAllODF

Tractography.ComputeAllODF — Type

struct ComputeAllODF <: Tractography.AbstractEvaluation

Spherical harmonics evaluation based on Fibonacci sampling. All ODF are pre-computed once. Their positivity is enforced with a max(0,⋅) or a mollifier.

Danger

Requires a relatively large memory!

Details

If you have na angles for sampling the unit sphere and the data is of size (nx,ny,nz,nsph), it yields a matrix of dimensions (nx,ny,nz,na).

See also FibonacciSH

Models

Tractography.TMC — Type

struct TMC{T, Talg<:Tractography.AbstractEvaluation, Td, TC, Tmol}

Tractography Markov Chain (TMC).

Fields (with default values):

Δt::Any: Step size of the TMC. Default: 0.1

evaluation_algo::Tractography.AbstractEvaluation: Spherical harmonics evaluation algorithm. Can be FibonacciSH(), ComputeAllODF(). Default: ComputeAllODF()

odfdata::Any: ODF data from nifti file. Must be the list of ODF in the base of spherical harmonics. Hence, it should be an (abstract) 4d array. Default: nothing

C::Any: Cone function to restrict angle diffusion. You can use a Cone or a custom function (d1,d2) -> return_a_boolean. Default: Cone(90)

proba_min::Any: Probability below which we stop tracking. Default: 0.0

mollifier::Any: Mollifier, used to make the fodf non negative. During odf evaluation, we effectively use mollifier(fodf[i,j,k,angle]). Default: default_mollifier

Methods

_apply_mask!(model, mask) apply a mask to the raw SH tensor. See its doc string.

_getdata(model) return the fodf data associated with the TMC.

size(model) return nx,ny,nz,nt.

eltype(model) return the scalar type of the data (default Float64).

get_lmax(model) return the max l coordinate in of spherical harmonics.

Constructors (use the fields!)

TMC()

TMC(Δt = 0.1f0) for a Float32 TMC

TMC(Δt = 0.1, proba_min = 0.) for a Float64 TMC. You need to specify both fields Δt and proba_min

TMC(odfdata = rand(10,10,10,45)) for custom ODF

Algorithms

Tractography.Deterministic — Type

struct Deterministic <: Tractography.DeterministicSampler

Tractography sampling performed with the argmax function. Can be used with FibonacciSH and ComputeAllODF.

Tractography.CSD — Type

struct CSD <: Tractography.AbstractNotPureRejectionSampler

Tractography sampling performed with the cumulative sum distribution. Can be used with FibonacciSH and ComputeAllODF.

Constructor

CSD()

Missing docstring.

Missing docstring for Diffusion. Check Documenter's build log for details.

Sampling

Tractography.init — Function

init(
    model::TMC{𝒯},
    alg;
    n_sphere,
    𝒯ₐ
) -> Tractography.ThreadedCache{_A, _B, _C, _D, _E, Nothing} where {_A, _B, _C, _D, _E}

Create a cache for computing streamlines in batches. This is interesting for use of memory limited environments (e.g. on GPU).

Tip

Use it with sample!

Arguments

alg sampling algorithm, Deterministic, CSD, Diffusion, etc.

n_sphere::Int = 400 number of points to discretize the sphere on which we evaluate the spherical harmonics.

𝒯ₐ = Array{𝒯} specify the type of the arrays in the cache. If passed a GPU array type like CuArray for CUDA, the sampling occurs on the GPU. Leave it for computing on CPU.

Tractography.sample — Function

sample(
    model::TMC{𝒯},
    alg::Tractography.AbstractSampler,
    seeds::AbstractArray{𝒯, 2};
    ...
) -> Tuple{Any, Any}
sample(
    model::TMC{𝒯},
    alg::Tractography.AbstractSampler,
    seeds::AbstractArray{𝒯, 2},
    mask::Union{Nothing, AbstractArray{Bool}};
    nt,
    n_sphere,
    maxodf_start,
    reverse_direction,
    nthreads,
    gputhreads
) -> Tuple{Any, Any}

Sample the TMC model.

Arguments

model::TMC

alg sampling algorithm, Deterministic, CSD, Diffusion, etc.

seeds matrix of size 6 x Nmc where Nmc is the number of Monte-Carlo simulations to be performed.

mask = nothing matrix of boolean where to stop computation. See also _apply_mask.

Optional arguments

nt::Int maximal length of each streamline.

n_sphere::Int = 400 number of points to discretize the sphere on which we evaluate the spherical harmonics.

maxodf_start::Bool for each locations, use direction provided by the argmax of the ODF.

reverse_direction::Bool reverse initial direction.

nthreads::Int = 8 number of threads on CPU.

gputhreads::Int = 512 number of threads on GPU.

Output

streamlines with shape 3 x nt x Nmc

Tractography.sample! — Function

sample!(
    streamlines,
    streamlines_length,
    model::TMC{𝒯},
    cache::Tractography.AbstractCache,
    alg,
    seeds;
    maxodf_start,
    reverse_direction,
    nthreads,
    gputhreads,
    nₜ,
    𝒯ₐ
) -> Any

Sample the TMC model inplace by overwriting result. This requires very little memory and can be run indefinitely on the GPU for example.

Arguments

streamlines array with shape 3 x nt x Nmc. nt is maximal length of each streamline. Nmc is the number of Monte-Carlo simulations to be performed.

streamlines_length length of the streamlines

model::TMC

alg sampling algorithm, Deterministic, CSD, Diffusion, etc.

seeds matrix of size 6 x Nmc where Nmc is the number of Monte-Carlo simulations to be performed.

Optional arguments

maxodf_start::Bool for each locations, use direction provided by the argmax of the ODF.

reverse_direction::Bool reverse initial direction.

nthreads::Int = 8 number of threads on CPU.

gputhreads::Int = 512 number of threads on GPU.

Plotting

Missing docstring.

Missing docstring for Tractography.plot_streamlines. Check Documenter's build log for details.

Tractography.plot_streamlines! — Function

plot_streamlines!(ax, streamlines; k...)

Plot the streamlines.

The optional parameters are passed to lines!

Tractography.plot_odf! — Function

plot_odf!(ax, model; n_sphere, radius, st, I, J, K, c0min)

Plot the the ODF with glyphs.

See also

plot_odf(model; kwargs...)

Arguments

model::TMC

I,J,K range for displaying the ODF. Some of them can be a single element, like K = 40:40.

radius radius of the glyph.

st = 4 stride, only show one over st glyph in each direction.

Optional arguments

Tractography.plot_slice — Function

plot_slice(model; k...)

Plot a slice of the data.

See also

plot_slice!

Arguments

model::TMC

Optional arguments

odf::Bool display the ODF with glyphs.

slice::Bool display the brain slice from mean ODF value.

interpolate::Bool interpolate the brain slice (reduces pixelization).

alpha alpha value for brain slice.

I,J,K range for displaying the ODF. One of them should be a single element, like K = 40:40.

st::Int = 2

c0min minimum mean ODF value for plotting the glyph.

n_theta::Int number of points for showing the glyphs.

radius = 0.1 radius of the glyphs.